chore: update versions

+++ src/initialization/node-version.ts
@@ -5,10 +5,14 @@ import chalk from 'chalk';
 export const createNodeVersionFiles = async () => {
   try {
     const packageJsonPath = path.join(process.cwd(), '.nvmrc');
-    fs.writeFileSync(packageJsonPath, 'v22');
+    if (!fs.existsSync(packageJsonPath)) {
+      fs.writeFileSync(packageJsonPath, 'v22');
+    }
 
     const nodeVersionPath = path.join(process.cwd(), '.node-version');
-    fs.writeFileSync(nodeVersionPath, 'v22');
+    if (!fs.existsSync(nodeVersionPath)) {
+      fs.writeFileSync(nodeVersionPath, 'v22');
+    }
   } catch (error) {
     console.log(chalk.red(error));
   }
